Names are now database objects rather than just strings.  They have a rank, a text name an author and notes.
A show_name page shows this information as well as a list of all
observations that use that Name.  The show_pages can be accessed from
each show_observation page as well as the name_index page that is
available in the left-side panel.

Names can be edited on the edit_name pages that are accessible from
the show_name pages.  Anyone is allowed to edit any name, but
following the Wikipedia model, a record is kept of every version.  The
previous versions are stored as 'past names'.  A show_past_name page
shows the name information for a previous version of a name and is
accessible from each show_name page.

Names are associated with observations through the 'what' field of the
observation.  When the 'what' field is changed, the names are searched
for a match.  The 'what' field can either be the text_name or the
text_name and the author for a name.  No match is found for a 'what'
value, then the user is given the option to create the new name.  If a
'what' value matches more than one name (due to different authors
publishing the same name), then the user is given a choice between the
options.  The options list how many observations use each version of
the name.  If you add an author to a 'what' field for a name that does
not have an author already, it will ask you if you want to create the
new name, but it will actually update the existing name.

Names given in species lists are treated in the exact same manner,
except that all the new names and all of the multiple matches are
dealt with on the same page.  There is also an option on the
new_species_list and edit_species_list pages to 'Just Create New
Names'.  This adds the names without updating the species list or
creating the corresponding observations.

The search box is now available on every page.

Removed the Observation Index in favor of the Name Index.

Fixed some of the list ordering to make it more intuitive.

A filter was added to the account creation process that should reduce the
number of bot created users.

Fixed a bug with bouncing between observations and images and using the
next/prev links.



git-svn-id: http://svn.collectivesource.com/mushroom_sightings/trunk@73 7a65b82a-1d14-0410-b525-e2835e71a0b4