Merged in a series of features added by Jason Hollinger.
Comments now visible directly on the Show Observation page. In
addition the Add Comment and Edit Comment pages now list all the
information for the observation to make it easier to cut and paste
while working on a comment.

Deprecating or approving a name now asks for a comment.

Better auto-completion enabled for Create Observation and Edit
Observation.


git-svn-id: http://svn.collectivesource.com/mushroom_sightings/trunk@193 7a65b82a-1d14-0410-b525-e2835e71a0b4