Fixed all the destroy_<object> I broke while writing the DELETE methods for the new api_controller. All tests pass now.
Changed textile wrapper to defer object lookup, and instead have _Name_, etc. link to observer/lookup_<type>/<id>.
This should save a considerable amount of database accesses.
The syntax for _Name_ has been simplified somewhat as a result: anything, essentially, that parses as a name (via Name.parse_name) is interpreted as a name, and linked (unmodified) to observer/lookup_name.
This means that things like _Lichens of North America_ are linked inappropriately (but harmlessly) to lookup_name.
But this also means that (except for the hypertext link) it looks exactly the way the user typed it now, so it should be safer.
If false-links become a problem, I suggest caching a list of all known genera, and only linkifying names that start with a valid genus.

API GET methods all seem to work well, but are not unit-tested.
API PUT, POST, and DELETE methods, while fully written (except for image upload) are completely untested.


git-svn-id: http://svn.collectivesource.com/mushroom_sightings/trunk@711 7a65b82a-1d14-0410-b525-e2835e71a0b4